Jos/requirements.txt
# Joy V1 Sales Rep - Voco V2 Sales Automation

# Orchestration
langgraph>=0.0.30
langchain-core>=0.1.0

# ML inference (Apple Silicon)
mlx>=0.0.1
mlx-lm>=0.0.1

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# X/Twitter scraping
snscrape>=0.7.0

# Environment variables
python-dotenv>=1.0.0

# Gmail API (OAuth2 fallback for email sending)
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0

# Web dashboard
flask>=2.0.0

# Scheduler
apscheduler>=3.10.0
pytz>=2024.1

# X/Twitter posting (OAuth 1.0a)
tweepy>=4.14.0

# LinkedIn engagement (unofficial API)
linkedin-api>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
